+++ android/app/src/main/java/com/yagubogu/data/datasource/auth/AuthDataSource.kt
@@ -1,6 +1,6 @@
-package com.yagubogu.data.datasource
+package com.yagubogu.data.datasource.auth
 
-import com.yagubogu.data.dto.response.LoginResponse
+import com.yagubogu.data.dto.response.auth.LoginResponse
 
 interface AuthDataSource {
     suspend fun login(idToken: String): Result<LoginResponse>
