+++ android/app/src/main/java/com/yagubogu/data/dto/request/auth/LogoutRequest.kt
@@ -1,4 +1,4 @@
-package com.yagubogu.data.dto.request
+package com.yagubogu.data.dto.request.auth
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
