+++ android/app/src/main/java/com/yagubogu/data/service/AuthApiService.kt
@@ -1,8 +1,8 @@
 package com.yagubogu.data.service
 
-import com.yagubogu.data.dto.request.LoginRequest
-import com.yagubogu.data.dto.request.LogoutRequest
-import com.yagubogu.data.dto.response.LoginResponse
+import com.yagubogu.data.dto.request.auth.LoginRequest
+import com.yagubogu.data.dto.request.auth.LogoutRequest
+import com.yagubogu.data.dto.response.auth.LoginResponse
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
